requirements.txt
# Core dependencies
fastapi>=0.115.0
uvicorn[standard]>=0.34.0
streamlit>=1.41.0
anthropic>=0.42.0
openai>=1.58.0
hume[microphone]>=0.7.0
python-dotenv>=1.0.1
websockets>=14.0
httpx>=0.28.0
pydantic>=2.10.0
gTTS>=2.5.0
SpeechRecognition>=3.10.0
